requirements.txt
pycryptodome
ecdsa